// src/types.ts
export interface InfactorySDKError {
  status: number;
  message: string;
}
export interface ApiResponse<T> {
  data?: T;
  error?: InfactorySDKError;
}

// Base types
export interface BaseEntity {
  id: string;
  created_at: string;
  updated_at: string;
  deleted_at?: string | null;
}
// ...
export interface QueryProgram extends BaseEntity {
  name?: string;
  query?: string;
  query_program?: string;
  steps?: string;
  slots?: string;
  stores?: string;
  published?: boolean;
  reason?: string;
  prev_id?: string;
  project_id: string;
}
// ...
export interface API extends BaseEntity {
  name: string;
  description?: string;
  project_id: string;
  base_path: string;
  version: string;
  tags?: string[];
  status: 'draft' | 'published' | 'deprecated';
  specification?: Record<string, any>;
}
export interface CreateAPIParams {
  name: string;
  project_id: string;
  base_path: string;
  version: string;
  description?: string;
  servers?: string;
  security?: string;
  tags?: string;
}

export interface EndpointTypes {
  GET: 'GET';
  POST: 'POST';
  PUT: 'PUT';
  DELETE: 'DELETE';
  PATCH: 'PATCH';
}

export interface APIEndpoint extends BaseEntity {
  name: string;
  description?: string;
  api_id: string;
  path: string;
  tags?: string;
  http_method: EndpointTypes[keyof EndpointTypes];
  request_schema?: Record<string, any>;
  response_schema?: Record<string, any>;
  implementation?: string;
  queryprogram_id?: string;
}

export interface CreateAPIEndpointParams {
  api_id: string;
  endpoint_name: string;
  http_method: EndpointTypes[keyof EndpointTypes];
  path: string;
  queryprogram_id: string;
  description?: string;
  operation_id?: string;
  tags?: string;
  parameters?: string;
  request_body?: string;
  responses?: string;
  security?: string;
}